Minor auth optimizations

+++ src/services/auth/src/main/kotlin/io/kuz/ecom/auth/app/service/SessionVerificationServiceImpl.kt
@@ -3,10 +3,13 @@ package io.kuz.ecom.auth.app.service
 import io.kuz.ecom.auth.app.service.extension.readChecked
 import io.kuz.ecom.auth.domain.VerificationCodeService
 import io.kuz.ecom.auth.domain.SessionVerificationService
+import io.kuz.ecom.auth.domain.exception.AlreadyExistsException
 import io.kuz.ecom.auth.domain.exception.InvalidCredentialsException
 import io.kuz.ecom.auth.domain.exception.TooEarlyException
+import io.kuz.ecom.auth.domain.model.AuthProvider
 import io.kuz.ecom.auth.domain.model.VerificationSessionData
 import io.kuz.ecom.auth.domain.model.VerificationVariant
+import io.kuz.ecom.auth.domain.repository.CredentialsRepository
 import io.kuz.ecom.auth.domain.repository.VerificationRepository
 import org.springframework.beans.factory.annotation.Value
 import org.springframework.stereotype.Service
@@ -15,6 +18,7 @@ import java.util.*
 
 @Service
 class SessionVerificationServiceImpl(
+    private val credentialsRepo: CredentialsRepository,
     private val verificationRepo: VerificationRepository,
     private val codeService: VerificationCodeService,
     @Value("\${app.verify.ttl}")
@@ -23,8 +27,21 @@ class SessionVerificationServiceImpl(
     private val verifySessionRetryTimeout: Long,
 ):SessionVerificationService {
 
-    // TODO: Add email validation
     override suspend fun initiateVerification(variant: VerificationVariant): VerificationSessionData {
+        val provider: AuthProvider
+        val identifier: String
+
+        when(variant) {
+            is VerificationVariant.Email -> {
+                provider = AuthProvider.LOCAL
+                identifier = variant.value
+            }
+        }
+
+        if (credentialsRepo.readRecord(provider, identifier) == null) {
+            throw AlreadyExistsException()
+        }
+
         verificationRepo.readVerificationSession(variant)?.let {
             verificationRepo.deleteVerificationSession(it.id)
         }
